# swagger.json
{
  "schemes": [],
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.",
    "title": "Swagger Example API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "api.postfly.kr",
  "basePath": "",
  "paths": {
...
    "/admin/v1/packing/bulk": {
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Packing"
        ],
        "summary": "Update the status of a Packing",
        "parameters": [
          {
            "description": "Request body",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/dto.UpdatePackingStatusRequest"
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/pkg.Response"
            }
          },
          "404": {
            "description": "Packing not found",
            "schema": {
              "$ref": "#/definitions/pkg.Response"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/pkg.Response"
            }
          }
        }
      }
    },
...
    "/admin/v1/users": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "User"
        ],
        "summary": "ListUsers",
        "parameters": [
          {
            "type": "string",
            "example": "customerNumber",
            "name": "customerNumber",
            "in": "query"
          },
          {
            "type": "string",
            "example": "John Doe",
            "description": "Filter by",
            "name": "name",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 1,
            "example": 1,
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 20,
            "example": 20,
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "응답값",
            "schema": {
              "$ref": "#/definitions/dto.ListUsersResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "domain.PackingStatus": {
      "type": "string",
      "enum": [
        "ADDING_ITEMS",
        "PACKING_REQUESTED",
        "PACKING_DONE",
        "PAYMENT_COMPLETED",
        "TRANSIT",
        "DELIVERED",
        "CHANGE_REQUESTED",
        "CANCELED"
      ],
      "x-enum-varnames": [
        "PackingStatusAddingItems",
        "PackingStatusPackingRequested",
        "PackingStatusPackingDone",
        "PackingStatusPaymentCompleted",
        "PackingStatusTransit",
        "PackingStatusDelivered",
        "PackingStatusChangeRequested",
        "PackingStatusCanceled"
      ]
    },
    "dto.AddressDTO": {
      "type": "object",
      "properties": {
        "address1": {
          "type": "string"
        },
        "address2": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "postcode": {
          "type": "string"
        },
        "state": {
          "type": "string"
        }
      }
    },
...
    "dto.GetUserInfoResponse": {
      "type": "object",
      "required": [
        "customerNumber",
        "email",
        "name",
        "userID"
      ],
      "properties": {
        "customerNumber": {
          "type": "string",
          "example": "customerNumber"
        },
        "defaultReceiverInfo": {
          "$ref": "#/definitions/dto.ReceiverInfoDTO"
        },
        "defaultSenderInfo": {
          "$ref": "#/definitions/dto.SenderInfoDTO"
        },
        "email": {
          "type": "string",
          "example": "email"
        },
        "name": {
          "type": "string",
          "example": "name"
        },
        "phoneNumber": {
          "type": "string",
          "example": "010-1234-5678"
        },
        "userID": {
          "type": "string",
          "example": "userID"
        }
      }
    },
...
    "dto.ListUsersResponse": {
      "type": "object",
      "required": [
        "users"
      ],
      "properties": {
        "page": {
          "type": "integer"
        },
        "size": {
          "type": "integer"
        },
        "total": {
          "type": "integer"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/dto.GetUserInfoResponse"
          }
        }
      }
    },
...
    "dto.ReceiverInfoDTO": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/dto.AddressDTO"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    },
    "dto.SenderInfoDTO": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    },
...
    "dto.UpdatePackingStatusRequest": {
      "type": "object",
      "required": [
        "packingID",
        "status"
      ],
      "properties": {
        "packingID": {
          "type": "string",
          "example": "packingID"
        },
        "status": {
          "allOf": [
            {
              "$ref": "#/definitions/domain.PackingStatus"
            }
          ],
          "example": "CREATED,REQUESTED,TRANSIT,DELIVERED"
        }
      }
    },
    "pkg.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {},
        "message": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}